Remove utils folder and refactor it in core directory
