+++ bigframes/core/__init__.py
@@ -25,6 +25,7 @@ import pandas
 
 import bigframes.core.compile as compiled
 import bigframes.core.guid
+from bigframes.core.log_adapter import class_logger, method_logger
 import bigframes.core.nodes as nodes
 from bigframes.core.ordering import OrderingColumnReference
 import bigframes.core.ordering as orderings
@@ -38,6 +39,8 @@ if typing.TYPE_CHECKING:
 
 ORDER_ID_COLUMN = "bigframes_ordering_id"
 PREDICATE_COLUMN = "bigframes_predicate"
+
+__all__ = ["class_logger", "method_logger"]
 
 
 @dataclass(frozen=True)
