+++ bigframes/operations/strings.py
@@ -18,11 +18,11 @@ import re
 from typing import cast, Literal, Optional, Union
 
 import bigframes.constants as constants
+from bigframes.core import log_adapter
 import bigframes.dataframe as df
 import bigframes.operations as ops
 import bigframes.operations.base
 import bigframes.series as series
-from bigframes.utils import log_adapter
 import third_party.bigframes_vendored.pandas.core.strings.accessor as vendorstr
 
 # Maps from python to re2
